fix(blevm): split lib proofs
